% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist.R
\name{rdist}
\alias{rdist}
\title{simulation d'une loi discrete}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{vecteur numerique represente les valeurs de l'aleatoire}

\item{p}{vecteur numerique des probabilites}
}
\description{
simulation d'une loi discrete
}
