% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdistn.R
\name{rdistn}
\alias{rdistn}
\title{simulation d'une loi discrete n fois}
\usage{
rdistn(x, p, n)
}
\arguments{
\item{x}{vecteur numerique represente une valeur d'une variable aleatoire}

\item{p}{vecteur numerique represente les probabilites}

\item{n}{nombre d'iteration}
}
\description{
simulation d'une loi discrete n fois
}
