% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMD.R
\name{scMD}
\alias{scMD}
\title{scMD framework}
\usage{
scMD(
  dat,
  bulk,
  nc = 5,
  DM_df = NULL,
  dmet_list = c("CIBERSORT", "EPIC", "FARDEEP", "DCQ", "DeconRNASeq", "BayesPrism",
    "ICeDT", "dtangle", "hspe"),
  include_minfi = TRUE,
  nmrk = 100,
  celltype_ind = NULL,
  includeMval = T,
  gen_sig = T,
  output_path = "."
)
}
\arguments{
\item{dat}{A beta matrix containing the scDNAm data.}

\item{bulk}{A matrix of bulk data.}

\item{nc}{An integer specifying the number of cores for parallel computation. Default is 5.}

\item{DM_df}{A data frame containing information about differentially methylated CpG sites from scDNAm data.
It should include columns for TargetID and celltype_ind. The columns celltype_ind contain p-values for each cell type.}

\item{dmet_list}{A character vector specifying the list of deconvolution methods. Default is c("CIBERSORT", "EPIC", "FARDEEP", "DCQ",
"DeconRNASeq", "BayesPrism", "ICeDT", "dtangle", "hspe").}

\item{include_minfi}{A logical value indicating whether to include Minfi-based deconvolution. Default is TRUE.}

\item{nmrk}{Number of markers per cell type used in deconvolution. Default is 100.}

\item{celltype_ind}{A character vector specifying the cell types to be included in the scMD analysis. Default is NULL, which
includes all cell types from the input data.}

\item{includeMval}{A logical value indicating whether to include M values (beta values transformed to M values) for deconvolution.
Default is TRUE.}

\item{gen_sig}{A logical value indicating whether to generate signatures for deconvolution. Default is TRUE. When set to FALSE, it means
the input data is already preprocessed to only include markers for deconvolution.}

\item{output_path}{A character specifying the path to save the output results. Default is the current directory (".").}
}
\value{
If scMD converges, return:
\itemize{
   \item {scMD_p: matrix, scMD estiamted CTS fractions.}
   \item {phat_all: list, estimated cell-type fractions.}
   \item {sig: matriix, signatures used in the deconvolution.}
   }
}
\description{
This function is used to perform signature generation and ensemble deconvolution using the scMD algorithm.
}
\examples{
# Assuming 'dat',  'bulk', 'DM_df', and 'sig_all' are defined
scMD(dat = my_dat, bulk = my_bulk, DM_df = my_DM_df, sig_all = my_sig_all)

}
