% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatDisc.R
\name{SD}
\alias{SD}
\title{statistique discriptive}
\usage{
SD(x)
}
\arguments{
\item{x}{vecteur numerique representant les valeurs de la variable aleatoire}
}
\description{
statistique discriptive
}
