% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMD.R
\name{scMD}
\alias{scMD}
\title{scMD framework}
\usage{
scMD(
  bulk,
  nc = 5,
  bulk_type = "450k",
  use_sc = "Luo",
  dat = NULL,
  DM_df = NULL,
  dmet_list = c("CIBERSORT", "EPIC", "FARDEEP", "DCQ", "DeconRNASeq", "BayesPrism",
    "ICeDT", "dtangle", "hspe"),
  include_minfi = TRUE,
  nmrk = 100,
  celltype_ind = NULL,
  includeMval = T,
  gen_sig = T,
  output_path = "."
)
}
\arguments{
\item{bulk}{A matrix of bulk data.}

\item{nc}{An integer specifying the number of cores for parallel computation. Default is 5.}

\item{bulk_type}{A character specifying the technical platform of bulk data. Default is "450k". It can also be chosen from "450k", "850k", "WGBS", or "NULL".
When set to NULL, the users should provide `dat` and `DM_df`.}

\item{use_sc}{A character specifying the inbuilt scDNAm to be used. Default is "Luo". It can also be chosen from "Luo", "Tian", or "NULL".
When set to NULL, the users should provide `dat` and `DM_df`.}

\item{dat}{A beta matrix containing the scDNAm data. Default is NULL. When set to NULL, scMD will used inbuilt scDNAm data instead.}

\item{DM_df}{A data frame containing information about differentially methylated CpG sites from scDNAm data.
It should include columns for TargetID and celltype_ind. The columns celltype_ind contain p-values for each cell type.
Default is NULL. When set to NULL, scMD will used inbuilt scDNAm data instead or the provided `dat` is already preprocessed.}

\item{dmet_list}{A character vector specifying the list of deconvolution methods. Default is c("CIBERSORT", "EPIC", "FARDEEP", "DCQ",
"DeconRNASeq", "BayesPrism", "ICeDT", "dtangle", "hspe").}

\item{include_minfi}{A logical value indicating whether to include Minfi-based deconvolution. Default is TRUE.}

\item{nmrk}{Number of markers per cell type used in deconvolution. Default is 100.}

\item{celltype_ind}{A character vector specifying the cell types to be included in the scMD analysis. Default is NULL, which
includes all cell types from the input data.}

\item{includeMval}{A logical value indicating whether to include M values (beta values transformed to M values) for deconvolution.
Default is TRUE.}

\item{gen_sig}{A logical value indicating whether to generate signatures for deconvolution. Default is TRUE. When set to FALSE, it means
the input data is already preprocessed to only include markers for deconvolution.}

\item{output_path}{A character specifying the path to save the output results. Default is the current directory (".").}
}
\value{
If scMD converges, return:
\itemize{
   \item {scMD_p: matrix, scMD estiamted CTS fractions.}
   \item {phat_all: list, estimated cell-type fractions.}
   \item {sig: matriix, signatures used in the deconvolution.}
   }
}
\description{
This function is used to perform signature generation and ensemble deconvolution using the scMD algorithm.
}
\examples{
# Assuming 'my_bulk' and 'my_bulk_type' are defined
scMD(bulk = my_bulk, bulk_type = my_bulk_type)

}
